Use separate array to store alive values.
The program used to change the grid on the fly as alive values were
calculated meaning that the program worked in unexpected ways. The
alive values are now first stored in an array and then the grid is
modified only on a second pass. The first pass is used to calculate
and store alive values and the second pass changes the grid to the
next state.